fix: move addressesOf in core to getAllAddressesIntoABlock into network
