+++ packages/core/tests/utils/bloom/bloom.unit.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, test } from '@jest/globals';
-import { addressUtils, bloom, bloomUtils, Hex0x } from '../../../src';
+import { bloom, bloomUtils, Hex0x } from '../../../src';
 import {
-    bloomTestCases,
+    blockAddressesFixture,
     blooms,
+    bloomTestCases,
     invalidAddressBloomTestCases,
     validAddressBloomTestCases,
-    valueTypeBloomTestCases,
-    expandedBlockDetail
+    valueTypeBloomTestCases
 } from './fixture';
 import {
     InvalidBloomError,
@@ -19,41 +19,9 @@ import {
  * @group unit/utils-bloom
  */
 describe('utils/bloom', () => {
-    test('bloomUtils.addressesOf', () => {
-        const expected = [
-            '0x0000000000000000000000000000456e65726779',
-            '0x1a8abd6d5627eb26ad71c0c7ae5224cdc640faf3',
-            '0x1eef8963e1222417af4dac0d98553abddb4a76b5',
-            '0x23a46368e4acc7bb2fe0afeb054def51ec56aa74',
-            '0x45429a2255e7248e57fce99e7239aed3f84b7a53',
-            '0x576da7124c7bb65a692d95848276367e5a844d95',
-            '0x5db3c8a942333f6468176a870db36eef120a34dc',
-            '0x6298c7a54720febdefd741d0899d287c70954c68',
-            '0x95fe74d1ae072ee45bdb09879a157364e5341565',
-            '0x9a107a75cff525b033a3e53cadafe3d193b570ec',
-            '0xa416bdda32b00e218f08ace220bab512c863ff2f',
-            '0xb2c20a6de401003a671659b10629eb82ff254fb8',
-            '0xb7591602c0c9d525bc3a7cf3c729fd91b8bf5bf6',
-            '0xbeae4bef0121f11d269aedf6adb227259d4314ad'
-        ];
-        bloomUtils
-            .addressesOf(expandedBlockDetail)
-            .filter((address) => {
-                return addressUtils.isAddress(address); // Remove empty addresses.
-            })
-            .forEach((actual) => {
-                expect(expected.includes(actual)).toBeTruthy();
-            });
-    });
-
     test('boolUtils.filterOf - bit per key - default', () => {
-        const addresses = bloomUtils
-            .addressesOf(expandedBlockDetail)
-            .filter((address) => {
-                return addressUtils.isAddress(address);
-            });
-        const filter = bloomUtils.filterOf(addresses);
-        addresses.forEach((address) => {
+        const filter = bloomUtils.filterOf(blockAddressesFixture);
+        blockAddressesFixture.forEach((address) => {
             expect(
                 bloomUtils.isAddressInBloom(
                     filter,
@@ -66,13 +34,8 @@ describe('utils/bloom', () => {
 
     test('boolUtils.filterOf -  bit per key - set', () => {
         const k = 16;
-        const addresses = bloomUtils
-            .addressesOf(expandedBlockDetail)
-            .filter((address) => {
-                return addressUtils.isAddress(address);
-            });
-        const filter = bloomUtils.filterOf(addresses, k);
-        addresses.forEach((address) => {
+        const filter = bloomUtils.filterOf(blockAddressesFixture, k);
+        blockAddressesFixture.forEach((address) => {
             expect(
                 bloomUtils.isAddressInBloom(
                     filter,
